requirements.txt
requests==2.31.0
docker==6.1.3
websocket-client==1.6.4
ruamel.yaml>=0.17.35
pytest==7.4.2
PyYAML~=6.0.1
setuptools~=68.2.2